Crate log_device

Structs

LogDevice

LogDeviceManager

StripeInfo

StripeReference

Here is the folder structure of log devices. path_dir/meta_db/ /log_device_0/segment_0 /segment_1 /… /log_device_1/segment_0 /segment_1 /… /… There is an independent file for each segment in a log. This is to facilitate log trimming in garbage collection.